4-python-Bunsans/pyproject.toml
[tool.poetry]
name = "backend-academy-2024-python-template"
version = "0.1.0"
description = ""
package-mode = false
authors = ["n.beskodarov <n.beskodarov@tinkoff.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.3"
tqdm = "^4.67.0"
loguru = "^0.7.2"
pillow = "^11.0.0"


[tool.poetry.dev-dependencies]
black = "^24.8.0"
isort = "^5.13.2"
ruff = "^0.6.1"
pytest = "*"
pytest-cov = "*"
pytest-deadfixtures = "*"
pytest-mock = "*"
pytest-repeat = "*"
pytest-asyncio = "*"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
